feat(common): Create shared utility package to reduce code duplication
Create chain/pkg/common/ with production-quality utilities:

1. validation.go - Address validation functions
   - ValidateAddress: Validates bech32 addresses and returns typed AccAddress
   - ValidateAddresses: Batch validation for multiple addresses
   - MustValidateAddress: Panic-on-error for guaranteed contexts

2. pagination.go - Pagination helpers
   - NormalizePagination: Ensures sensible defaults (100/page, max 1000)
   - GetEffectiveLimit: Returns actual limit after normalization
   - DoS protection via bounded queries

3. store.go - Generic KVStore operations
   - GetObject: Type-safe object retrieval with clear not-found vs error states
   - SetObject: Type-safe object storage with marshal error handling
   - DeleteObject: Consistent delete semantics
   - HasObject: Efficient existence checks without deserialization
   - IterateObjects: Safe iteration with proper cleanup

4. events.go - Event emission helpers
   - EmitEvent: Consistent event emission across modules
   - EmitTypedEvent: Protobuf typed event support
   - EmitSuccessEvent: Standard success event structure
   - EmitErrorEvent: Standard error event structure
   - EmitTransferEvent: Token transfer event helper

5. errors.go - Shared error definitions
   - Common error codes (InvalidAddress, NotFound, Unauthorized, etc.)
   - Registered errors for consistent error handling
   - WrapError functions for adding context
   - IsXXXError helpers for error type checking

All code:
- Compiles successfully
- Has comprehensive test coverage (all tests pass)
- Includes security considerations in documentation
- Follows Cosmos SDK patterns
- Ready for module integration

Addresses ROADMAP task P3-117.

Next step: Update identity, dex, bridge, compliance, vcregistry modules to use these utilities.

+++ chain/pkg/common/pagination_test.go
@@ -115,9 +115,9 @@ func TestGetEffectiveLimit(t *testing.T) {
 }
 
 func TestDefaultPageLimit(t *testing.T) {
-	require.Equal(t, uint64(100), common.DefaultPageLimit)
+	require.Equal(t, uint64(100), uint64(common.DefaultPageLimit))
 }
 
 func TestMaxPageLimit(t *testing.T) {
-	require.Equal(t, uint64(1000), common.MaxPageLimit)
+	require.Equal(t, uint64(1000), uint64(common.MaxPageLimit))
 }
